create a separate py file for the decorator
