+++ launch_testing/launch_testing/__init__.py
@@ -15,7 +15,8 @@
 from . import tools
 from .decorator import post_shutdown_test
 from .io_handler import ActiveIoHandler, IoHandler
-from .parametrize import parametrize, ready_to_test_action_timeout
+from .parametrize import parametrize
+from .ready_to_test_action_timeout import ready_to_test_action_timeout
 from .proc_info_handler import ActiveProcInfoHandler, ProcInfoHandler
 from .ready_aggregator import ReadyAggregator
