+++ launch_testing/launch_testing/parametrize.py
@@ -55,17 +55,4 @@ def parametrize(argnames, argvalues):
     return _decorator
 
 
-def ready_to_test_action_timeout(timeout):
-    """
-    Decorate a test launch description in a way that it adds ReadyToTest action timeout
 
-    attribute to the function being decorated.
-
-    :param: timeout Duration for which the ReadyToTest action waits for processes to start up
-
-    """
-    def _decorator(func):
-        func.__ready_to_test_action_timeout__ = timeout
-        return func
-
-    return _decorator
